refactor: 코드리뷰를 반영하여 리팩토링
